Display the git revision when running from a git clone.
My proposed implementation admittedly is reaching out deep into LMS code... but it should allow us to reliably figure out where the plugin is being installed, as it can be in quite a few different places. It does get the list of folders being used by LMS' templating engine, and picks the one matching the MaterialSkin name. That folder can be expected to be in the git clone, where we can get the revision.

Fall back to timestamp if git fails.