Expand names and descriptions in web app manifests
I find three letter acronyms unfriendly.  There's not enough space in the `short_name` field to spell the name out in full, but there is in the `name` field.  I realise this is subjective, so won't be offended if this is rejected.

While here I've added a category and a description.

I've also added a `.editorconfig` file which will cause compatible editors to indent .webmanifest files by 2 columns.