Enable a limited set of the Interface Settings
Disabling this page has lead to some irritation - and cranky reactions. I believe that it should not be removed all together, as it covers some preferences which are not only crucial to the default web UI, but are settings to be used on the old players, too (Boom, Classic etc.).

I believe removing the skin selector alone has led to more confusion than leaving the full page there would ever have caused. I'd therefore suggest you include a limited set of options to ease the pain both for your users, as well as yourself. This patch does hide the options I believe you consider confusing, but still gives users access to UI settings they might need, like available formatting options in the classic player UI, the skin selector etc. This choice obviously could be tweaked if desired.

What I'm doing here is copy the original template and replace the input fields with hidden fields and put it in your skin folder. This form hasn't changed in years. This should be pretty safe.