Use LMS' own advanced search params parser (if available - requires 8.2+)
If LMS supports the `Slim::Web::Pages::Search::parseAdvancedSearchParams()` function, use it instead of the plugin's copy of that code.

In order to achieve this I've added a stub inline module `Plugins::MaterialSkin::Search`. If LMS doesn't support the aforementioned call, the plugin is requiring the `Search.pm` module, which would override the stub with its own implementation.