czo-sync-menu
Put 2 options : "Show the manage players menu" and "Show the synchronise menu"
It's like always for people who don't act on these two options. But for me, it's great: I can set "Show management drives" to OFF and "Show sync menu" to ON.
I have a lot of squeezeboxes, 1 Transporter, 1 Classic, 1 Boom, 3 Radio and my 2 laptops that use squeezelite.
I'm 52, and presbyopic, so on my phone, with a large font size, it's really hard to see all my squeezeboxes in the "manage drives" dialog, so I prefer to have the synchronise drop down menu for that.